The description was misleading
It states that creating a new array by shifting all the zeros to the left. I used a simple Sort Method will accomplish that. Then on the solution page, it states the same thing with the added condition "by swapping with number". You're supposed to swap positions of the zero and the left most number. I wasn't aware of this, so my solution didn't match theirs at first.